+++ android/vendored/unversioned/react-native-pager-view/android/src/main/java/com/reactnativepagerview/ViewPagerViewHolder.kt
@@ -1,4 +1,4 @@
-package versioned.host.exp.exponent.modules.api.components.pagerview
+package com.reactnativepagerview
 
 import android.view.ViewGroup
 import android.view.ViewGroup.LayoutParams.MATCH_PARENT